backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
ultralytics
opencv-python-headless
azure-storage-blob
python-dotenv
aiofiles
pillow
